// rss-css-selectors/src/components/controller/enum.ts
enum STYLES {
  HtmlSelectorFocus = 'selector--focus',
  PlayItemFocus = 'table__item--focus',
  PopupItemActive = 'popup--active',
  ListItemSelected = 'levels__item--selected',
  InputFilled = 'input--filled',
  LevelStatusActive = 'level-status--active',
  HintStatusActive = 'hint-status--active',
  LevelsOpen = 'levels--open',
  BurgerActive = 'btn-wrapper--active',
  ButtonEnterActive = 'btn--active',
  TableItemOut = 'table--out',
  PopupOpen = 'popup-container--open',
}

enum EVENTS {
  Mouseover = 'mouseover',
  Mouseout = 'mouseout',
  Click = 'click',
  Keydown = 'keydown',
}

enum KEYNAME {
  Enter = 'Enter',
}

export { STYLES, EVENTS, KEYNAME };