// rss-css-selectors/src/types/controller.ts
interface IViewElements {
  htmlSelectors: HTMLElement[];
  playElements: HTMLElement[];
  popupElements: HTMLElement[];
  buttonsControl: HTMLElement[];
  inputButtons: HTMLElement[];
  levelsList: HTMLElement[];
  inputField: HTMLElement;
  sidebarTitle: HTMLElement;
  levelsPanel: HTMLElement;
  tableField: HTMLElement;
  popup: HTMLElement;
}

export default IViewElements;